Revert "Merge pull request #48 from ignacioserranogonzalez/Task_BoolStringFloat"
This reverts commit 428f9b6751b4a1c637e97f7301c7ec6fe0572642, reversing
changes made to a420716896174982cf44752fa6ece35c0648938e.

+++ app/src/main/java/una/paradigmas/ast/FloatLiteral.java
@@ -13,7 +13,8 @@ package una.paradigmas.ast;
  * 
  * Codigo de grupo: 02-1PM
  */
-public record FloatLiteral(double value) implements Node {
+
+public record FloatLiteral(float value) implements Node {
     @Override
     public <T> T accept(Visitor<T> visitor) {
         return visitor.visitFloat(this);
